requirements.txt
streamlit==1.40.2
youtube-transcript-api==0.6.2
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
openai==1.35.0
networkx==3.2.1
plotly==5.17.0
pandas==2.1.4
selenium==4.15.2
beautifulsoup4==4.12.2